update: revise app menu code and implement settings and about functionality
